[REF] Separa classes do contrato em arquivos dedicados
