+++ l10n_br_hr_contract/models/hr_contract.py
@@ -23,99 +23,6 @@ MONTHS = [
 ]
 
 
-class HrContractAdmissionType(models.Model):
-    _name = 'hr.contract.admission.type'
-
-    name = fields.Char(string='Admission type')
-    code = fields.Char(string='Code')
-
-    @api.multi
-    def name_get(self):
-        result = []
-        for record in self:
-            name = record['name']
-            if record['code']:
-                name = record['code'] + ' - ' + name
-            result.append((record['id'], name))
-        return result
-
-
-class HrContractLaborBondType(models.Model):
-    _name = 'hr.contract.labor.bond.type'
-
-    name = fields.Char(string='Labor bond')
-    code = fields.Char(string='Code')
-
-    @api.multi
-    def name_get(self):
-        result = []
-        for record in self:
-            name = record['name']
-            if record['code']:
-                name = record['code'] + ' - ' + name
-            result.append((record['id'], name))
-        return result
-
-
-class HrContractLaborRegime(models.Model):
-    _name = 'hr.contract.labor.regime'
-
-    name = fields.Char(string='Labor regime')
-    short_name = fields.Char(string='Short name')
-    code = fields.Char(string='Code', size=1)
-
-    @api.multi
-    def name_get(self):
-        result = []
-        for record in self:
-            name = record['name']
-            if record['short_name']:
-                name = record['short_name'] + ' - ' + name
-            result.append((record['id'], name))
-        return result
-
-
-class HrContractSalaryUnit(models.Model):
-    _name = 'hr.contract.salary.unit'
-
-    name = fields.Char(string='Salary unit')
-    code = fields.Char(string='Code')
-
-    @api.multi
-    def name_get(self):
-        result = []
-        for record in self:
-            # name = record['name']
-            # if record['code']:
-            #     name = record.code + '-' + record.name
-            result.append((record['id'], record.code + ' - ' + record.name))
-        return result
-
-
-class HrContractResignationCause(models.Model):
-    _name = 'hr.contract.resignation.cause'
-
-    name = fields.Char(string='Resignation cause')
-    code = fields.Char(string='Resignation cause code')
-    fgts_withdraw_code = fields.Char(string='FGTS withdrawal code')
-
-    @api.multi
-    def name_get(self):
-        result = []
-        for record in self:
-            name = record['name']
-            if record['code']:
-                name = record['code'] + ' - ' + name
-            result.append((record['id'], name))
-        return result
-
-
-class HrContractNoticeTermination(models.Model):
-    _name = 'hr.contract.notice.termination'
-
-    name = fields.Char(string='Notice of termination type')
-
-
 class HrContract(models.Model):
     _inherit = 'hr.contract'
