+++ l10n_br_hr_contract/models/hr_contract_admission_type.py
@@ -1,0 +1,22 @@
+# -*- coding: utf-8 -*-
+# Copyright (C) 2016  Daniel Sadamo - KMEE Informática
+# License AGPL-3 - See http://www.gnu.org/licenses/agpl-3.0.html
+
+from odoo import models, fields, api
+
+
+class HrContractAdmissionType(models.Model):
+    _name = 'hr.contract.admission.type'
+
+    name = fields.Char(string='Admission type')
+    code = fields.Char(string='Code')
+
+    @api.multi
+    def name_get(self):
+        result = []
+        for record in self:
+            name = record['name']
+            if record['code']:
+                name = record['code'] + ' - ' + name
+            result.append((record['id'], name))
+        return result
